requirements.txt
streamlit>=1.28.0
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
python-dateutil>=2.8.0
